refactor: move profile manager to service package and rename in ProfileService
