+++ src/main/scala/users_management/service/ProfileService.scala
@@ -1,7 +1,8 @@
-package users_management.model
+package users_management.service
 
+import users_management.model.{Profile, UserProfile}
 
-trait ProfileManager:
+trait ProfileService:
   
   def getProfile(id: String): Option[Profile]
 
@@ -10,6 +11,8 @@ trait ProfileManager:
   def updateProfile(userProfile: UserProfile): Option[Profile]
 
   def deleteProfile(id: String): Boolean
+  
+  def deleteAllProfiles(): Boolean
 
   def listProfiles(): List[Profile]
